Split scanner orchestration into modules
